Service/Task update for debugging
